Added a Matlab App to run basic simulations
The app can help to configure simulation cases and run them. It can show a preview of the Layout, the wind direction and the wind shear. It is meant to run the farm conners experiments. Storage is not  implemented yet and the yaw settings only work for the farm connors layout as intended.