Changes in Plotting
Added a feature that enforces full wake interaction for the last timestep IF the flow field is being plotted. This way the user can enjoy the speed of the reduced interaction and get the full flow field, as expected. The flow field is also exported as .vtk file as well as .fig IF the storing is activated. If the app is used, the files are stored in one folder with the generated power .csv. New colormaps are now also available such as viridis - needs to be activated by code or in the App.
Values are added (as comments) from Bart Doekemeijer's FLORIS Identification - they differ quite a lot from the used values which underlines the need of an Identification process.